[ALPS07799868][WCNR00379700] fix overflow sink coverity
[Description]
fix overflow sink coverity CID:#5249237

[Test]
UT

MTK-Commit-Id: a12d0e26f7e097f2cb5432cccb797a4020cbaae7

CR-Id: ALPS07799868
Feature: Wi-Fi Driver CONNAC
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/9137969
Test: srv_mspautosanity
CodeQL-Review-Label: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Commit-Check: srv_check_service <srv_check_service@mediatek.com>
ODB-Check: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
AutoUT-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_pf_nep_sanity
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Tested-by: jo-ping.li <jo-ping.li@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>

+++ include/chips/fw_dl.h
@@ -145,7 +145,7 @@ struct patch_dl_buf {
 
 struct patch_dl_target {
 	struct patch_dl_buf *patch_region;
-	uint8_t num_of_region;
+	uint32_t num_of_region;
 };
 
 struct WIFI_VER_INFO;
